pyproject.toml
[project]
name = "coom"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click>=8.3.0",
    "datasets>=4.1.1",
    "lightning>=2.5.5",
    "megatron-core==0.15.0rc6",
    "packaging>=25.0",
    "pyyaml>=6.0.3",
    "regex>=2025.9.18",
    "six>=1.17.0",
    "transformers>=4.56.2",
]

[tool.uv]
dev-dependencies = [
    "coom",
    "pytest>=8.3.4",
]

[tool.uv.sources]
coom = { workspace = true }

[tool.pytest.ini_options]
pythonpath = ["."]
addopts = ["--import-mode=importlib", ]
